Created basic UI structure
I have opted to ignore window resizing. Rather, the only "resizing" will be restricted to the default window size (1600 x 1000px), and full screen. For now, this allows me to isolate my efforts towards the functionality of the games and creating a consistent UI. This can be changed later as the project continues.