Added full list of games
Added the multiplayer games. Finished organization of menu items. Seems to be a limitation for the max size Visual Studio allows for the window. I am limited to 1460, as the monitor I'm developing on is 1440p. Controls can still be placed beyond these limits, but once they are out of view in the Designer, their location will need to be changed in code alone.